requirements.txt
opencv-python-headless
imagehash 
Pillow 
tqdm